+++ src/app/api/auth/callback/route.ts
@@ -39,6 +39,7 @@ export async function GET(request: NextRequest) {
       path: '/',
       maxAge: 60 * 60 * 24 * 365,
       sameSite: 'lax',
+      secure: true,
     })
   }